Split legacy views and make v4 the current one.
