+++ routes/index.js
@@ -67,10 +67,7 @@ function bootswatch(req, res) {
 
 function bootswatch4(req, res) {
     res = appendLocals(req, res);
-    res.render('bootswatch4', {
-        title: 'Bootswatch 4 Beta',
-        description: 'The recommended CDN for Bootswatch 4 Beta'
-    });
+    res.redirect(301, '/bootswatch/');
 }
 
 function bootlint(req, res) {
@@ -81,24 +78,29 @@ function bootlint(req, res) {
     });
 }
 
-function alpha(req, res) {
+function redirectToRoot(req, res) {
     res = appendLocals(req, res);
-    res.redirect(301, '/beta/');
-}
-
-function beta(req, res) {
-    res = appendLocals(req, res);
-    res.render('beta', {
-        title: 'Bootstrap 4 Beta',
-        description: 'The recommended CDN for Bootstrap 4 Beta'
-    });
+    res.redirect(301, '/');
 }
 
 function legacy(req, res) {
     res = appendLocals(req, res);
-    res.render('legacy', {
+    res.redirect(301, '/legacy/bootstrap/');
+}
+
+function legacyBootstrap(req, res) {
+    res = appendLocals(req, res);
+    res.render('legacy/bootstrap.pug', {
         title: 'Bootstrap Legacy',
         description: 'Older versions of Bootstrap hosted on a CDN'
+    });
+}
+
+function legacyBootswatch(req, res) {
+    res = appendLocals(req, res);
+    res.render('legacy/bootswatch.pug', {
+        title: 'Bootswatch Legacy',
+        description: 'Older versions of Bootswatch hosted on a CDN'
     });
 }
 
@@ -156,9 +158,10 @@ module.exports = {
     bootswatch,
     bootswatch4,
     bootlint,
-    alpha,
-    beta,
+    redirectToRoot,
     legacy,
+    legacyBootstrap,
+    legacyBootswatch,
     showcase,
     integrations
 };
