+++ tests/bootswatch_test.js
@@ -4,7 +4,7 @@ const path     = require('path');
 const assert   = require('assert');
 const helpers  = require(path.join(__dirname, 'test_helper.js'));
 const config   = helpers.config();
-const uri      = helpers.app(config, 'bootswatch');
+const uri      = helpers.app(config, 'legacy/bootswatch');
 
 let response = {};
