+++ tests/index_test.js
@@ -15,6 +15,91 @@ before((done) => {
     });
 });
 
+describe('bootstrap4', () => {
+    const latest = config.bootstrap4[0];
+
+    describe('config', () => {
+        it('is latest', (done) => {
+            assert(latest.latest);
+            done();
+        });
+
+        it('has stylesheet integrity', (done) => {
+            assert(typeof latest.stylesheetSri !== 'undefined');
+            done();
+        });
+
+        it('has javascript integrity', (done) => {
+            assert(typeof latest.javascriptSri !== 'undefined');
+            done();
+        });
+
+        it('has javascript bundle integrity', (done) => {
+            assert(typeof latest.javascriptBundleSri !== 'undefined');
+            done();
+        });
+    });
+
+    it('works', (done) => {
+        helpers.assert.response(response);
+        done();
+    });
+
+    it('contains authors', (done) => {
+        config.authors.forEach((author) => {
+            helpers.assert.contains(author, response.body);
+        });
+        done();
+    });
+
+    //it('has header', (done) => {
+    //    helpers.assert.contains('<h2 class="text-center mb-4">Bootstrap 4 Beta</h2>', response.body);
+    //    done();
+    //});
+
+    it('has notice', (done) => {
+        helpers.assert.contains('Bootstrap 4 is currently in Beta release and should be treated as such.', response.body);
+        done();
+    });
+
+    describe('stylesheet', () => {
+        it('has uri', (done) => {
+            helpers.assert.contains(latest.stylesheet, response.body);
+            done();
+        });
+
+        ['html', 'pug', 'haml'].forEach((fmt) => {
+            it(`has ${fmt}`, (done) => {
+                const str = helpers.css[fmt](latest.stylesheet, latest.stylesheetSri);
+
+                helpers.assert.contains(str, response.body);
+                done();
+            });
+        });
+    });
+
+    describe('javascript', () => {
+        it('has javascript uri', (done) => {
+            helpers.assert.contains(latest.javascript, response.body);
+            done();
+        });
+
+        it('has javascript bundle uri', (done) => {
+            helpers.assert.contains(latest.javascriptBundle, response.body);
+            done();
+        });
+
+        ['html', 'pug', 'haml'].forEach((fmt) => {
+            it(`has ${fmt}`, (done) => {
+                const str = helpers.javascript[fmt](latest.javascript, latest.javascriptSri);
+
+                helpers.assert.contains(str, response.body);
+                done();
+            });
+        });
+    });
+});
+
 describe('index', () => {
     const latest = config.bootstrap[0];
