+++ views/_partials/beta-alert.pug
@@ -1,4 +1,4 @@
-.card.border-info
+.card.border-info.my-4
     .card-header.bg-info.text-white
         h3.card-title Info
     .card-body
