--- views/bootswatch4.pug
@@ -1,35 +1,0 @@
-extends layout.pug
-
-block content
-    h2.text-center.mb-4 Bootswatch 4 Beta
-
-    include _partials/beta-alert.pug
-
-    -for (var i = 0, len = config.bootswatch4.themes.length; i < len; i++)
-        -var name = config.bootswatch4.themes[i].name.toLowerCase();
-        -var sri  = config.bootswatch4.themes[i].sri;
-        -var file = config.bootswatch4.bootstrap.replace('SWATCH_NAME', name).replace('SWATCH_VERSION', config.bootswatch4.version);
-
-        .card.bg-light.my-4.p-4.text-center
-            .row
-                .col-md-2
-                    .tryit.mt-md-4
-                        if (i === helpers.theme.selected(config, theme))
-                            button.btn.btn-primary(type='button', disabled) Active!
-                        else
-                            a.btn.btn-primary(role='button', href='?theme=' + i) Try it!
-
-                .col-sm-12.col-md-8
-                    .mt-4
-                        a(href=config.bootswatch4.link.replace('SWATCH_NAME', name), target='_blank', rel='noopener')
-                            img.bootswatch.d-block.img-fluid.mx-auto(src=config.bootswatch4.image.replace('SWATCH_NAME', name), alt=name + ' thumbnail', width='528', height='317')
-                        .input-group.mt-3
-                            input.form-control(type='text', value=file)
-                            span.input-group-btn
-                                button.btn.btn-secondary(type='button', data-toggle='collapse', data-target='#' + name, aria-expanded='false', aria-controls=name)
-                                    span.caret
-                        span.form-text.text-muted.mb-4 Click to copy
-
-                        include _partials/csscode.pug
-
-//- vim: ft=pug sw=4 sts=4 et:
