+++ views/index.pug
@@ -2,42 +2,6 @@ extends layout.pug
 
 block content
     h2.text-center=title
-    h3.text-center=`v${config.bootstrap[0].version}`
-
-    .card.bg-light.mt-4.p-4.text-center
-        -var name   = 'quickstartCSS'.toLowerCase() + config.bootstrap[0].version.replace(/\./g, '_');
-        -var file   = config.bootstrap[0].stylesheet;
-        -var sri    = config.bootstrap[0].stylesheetSri;
-        -var formId = name + '_form';
-
-        .form-group.my-4
-            label.font-weight-bold(for=formId) Complete CSS
-            .input-group.input-group-lg
-                input.form-control(id=formId, type='text', value=file)
-                span.input-group-btn
-                    button.btn.btn-secondary(type='button', data-toggle='collapse', data-target='#' + name, aria-expanded='false', aria-controls=name)
-                        span.caret
-            span.form-text.text-muted Click to copy
-
-        include _partials/csscode.pug
-
-        -var name   = 'quickstartJS'.toLowerCase() + config.bootstrap[0].version.replace(/\./g, '_');
-        -var file   = config.bootstrap[0].javascript;
-        -var sri    = config.bootstrap[0].javascriptSri;
-        -var formId = name + '_form';
-
-        .form-group.my-4
-            label.font-weight-bold(for=formId) Complete JavaScript
-            .input-group.input-group-lg
-                input.form-control(id=formId, type='text', value=file)
-                span.input-group-btn
-                    button.btn.btn-secondary(type='button', data-toggle='collapse', data-target='#' + name, aria-expanded='false', aria-controls=name)
-                        span.caret
-            span.form-text.text-muted Click to copy
-
-        include _partials/jscode.pug
-
-    hr.my-4
 
     include _partials/beta-alert.pug
 
@@ -83,6 +47,43 @@ block content
 
         .form-group.my-4
             label.font-weight-bold(for=formId) Complete JavaScript Bundle
+            .input-group.input-group-lg
+                input.form-control(id=formId, type='text', value=file)
+                span.input-group-btn
+                    button.btn.btn-secondary(type='button', data-toggle='collapse', data-target='#' + name, aria-expanded='false', aria-controls=name)
+                        span.caret
+            span.form-text.text-muted Click to copy
+
+        include _partials/jscode.pug
+
+    hr.my-4
+
+    h3.text-center=`v${config.bootstrap[0].version}`
+
+    .card.bg-light.mt-4.p-4.text-center
+        -var name   = 'quickstartCSS'.toLowerCase() + config.bootstrap[0].version.replace(/\./g, '_');
+        -var file   = config.bootstrap[0].stylesheet;
+        -var sri    = config.bootstrap[0].stylesheetSri;
+        -var formId = name + '_form';
+
+        .form-group.my-4
+            label.font-weight-bold(for=formId) Complete CSS
+            .input-group.input-group-lg
+                input.form-control(id=formId, type='text', value=file)
+                span.input-group-btn
+                    button.btn.btn-secondary(type='button', data-toggle='collapse', data-target='#' + name, aria-expanded='false', aria-controls=name)
+                        span.caret
+            span.form-text.text-muted Click to copy
+
+        include _partials/csscode.pug
+
+        -var name   = 'quickstartJS'.toLowerCase() + config.bootstrap[0].version.replace(/\./g, '_');
+        -var file   = config.bootstrap[0].javascript;
+        -var sri    = config.bootstrap[0].javascriptSri;
+        -var formId = name + '_form';
+
+        .form-group.my-4
+            label.font-weight-bold(for=formId) Complete JavaScript
             .input-group.input-group-lg
                 input.form-control(id=formId, type='text', value=file)
                 span.input-group-btn
